authorizing who can edit and delete

+++ app/Http/Controllers/IdeaController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Idea;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class IdeaController extends Controller
 {
@@ -20,6 +21,8 @@ class IdeaController extends Controller
             'content' => 'required|min:3|max:240'
         ]);
 
+        $validated['user_id'] = Auth::id();
+
         Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Idea was created successfully!');
@@ -32,6 +35,10 @@ class IdeaController extends Controller
     */
     public function destroy(Idea $idea)
     {
+        if (Auth::user()->id !== $idea->user_id) {
+            abort(404);
+        }
+
         try {
             $idea->delete();
         } catch (ModelNotFoundException $e) {
@@ -42,13 +49,25 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (Auth::user()->id !== $idea->user_id) {
+            abort(404);
+        }
+
         $editing = true;
 
         return view('ideas.show', compact('idea', 'editing'));
     }
 
+    public function canUpdate(Idea $idea)
+    {
+        return Auth::user()->id == $idea->user_id;
+    }
     public function update(Idea $idea)
     {
+        if (Auth::user()->id !== $idea->user_id) {
+            abort(404);
+        }
+
         $validated = request()->validate([
             'content' => 'required|min:3|max:240'
         ]);
